requirements.txt
Django~=5.0.4
djangorestframework~=3.15.1
python-dotenv~=1.0.1
django-cors-headers~=4.3.1
drf-yasg~=1.21.7
psycopg2~=2.9.9
gunicorn==22.0.0